move inertial_length() into lengths.py
