+++ plasmapy/formulary/parameters.py
@@ -50,12 +50,17 @@ import numbers
 import numpy as np
 import warnings
 
-from astropy.constants.si import c, e, eps0, k_B, mu0
+from astropy.constants.si import e, eps0, k_B, mu0
 from numba import njit
 from typing import Optional, Union
 
 from plasmapy import particles
-from plasmapy.formulary.lengths import Debye_length, lambdaD_  # noqa
+from plasmapy.formulary.lengths import (  # noqa
+    cwp_,
+    Debye_length,
+    inertial_length,
+    lambdaD_,
+)
 from plasmapy.particles import Particle
 from plasmapy.particles.exceptions import ChargeError
 from plasmapy.utils import PhysicsError
@@ -1739,83 +1744,6 @@ def Debye_number(T_e: u.K, n_e: u.m ** -3) -> u.dimensionless_unscaled:
 
 nD_ = Debye_number
 """Alias to `~plasmapy.formulary.parameters.Debye_number`."""
-
-
-@validate_quantities(
-    n={"can_be_negative": False},
-    validations_on_return={"equivalencies": u.dimensionless_angles()},
-)
-@particles.particle_input(require="charged")
-def inertial_length(n: u.m ** -3, particle: Particle) -> u.m:
-    r"""
-    Calculate a charged particle's inertial length.
-
-    **Aliases:** `cwp_`
-
-    Parameters
-    ----------
-    n : `~astropy.units.Quantity`
-        Particle number density in units convertible to m\ :sup:`-3`\ .
-
-    particle : `~plasmapy.particles.particle_class.Particle`
-        Representation of the particle species (e.g., 'p+' for protons,
-        'D+' for deuterium, or 'He-4 +1' for singly ionized helium-4).
-
-    Returns
-    -------
-    d : `~astropy.units.Quantity`
-        The particle's inertial length in meters.
-
-    Raises
-    ------
-    `TypeError`
-        If ``n`` is not a `~astropy.units.Quantity` or ``particle`` is
-        not a string.
-
-    `~astropy.units.UnitConversionError`
-        If ``n`` is not in units of a number density.
-
-    `ValueError`
-        The particle density does not have an appropriate value.
-
-    Warns
-    -----
-    : `~astropy.units.UnitsWarning`
-        If units are not provided and SI units are assumed.
-
-    Notes
-    -----
-    The inertial length of a particle of species :math:`s` is given by
-
-    .. math::
-        d = \frac{c}{ω_{ps}}
-
-    The inertial length is the characteristic length scale for a
-    particle to be accelerated in a plasma.  The Hall effect becomes
-    important on length scales shorter than the ion inertial length.
-
-    The inertial length is also known as the skin depth.
-
-    Examples
-    --------
-    >>> from astropy import units as u
-    >>> inertial_length(5 * u.m ** -3, 'He+')
-    <Quantity 2.02985...e+08 m>
-    >>> inertial_length(5 * u.m ** -3, 'e-')
-    <Quantity 2376534.75... m>
-
-    """
-    omega_p = plasma_frequency(n, particle=particle)
-
-    return c / omega_p
-
-
-cwp_ = inertial_length
-"""
-Alias to `~plasmapy.formulary.parameters.inertial_length`.
-
-* Name is shorthand for :math:`c / ω_p`.
-"""
 
 
 @validate_quantities
